requirements.txt
# Core web framework and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP client and async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
sqlite3  # Built-in, but explicit for clarity

# Authentication and security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data validation and serialization
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Caching and session management
redis>=5.0.0
aioredis>=2.0.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.4
prometheus-client>=0.19.0

# API integrations
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ML and data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Testing (for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.25.0  # For FastAPI testing

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Deployment
gunicorn>=21.2.0
docker>=6.1.0

# Environment and configuration
python-dateutil>=2.8.0
pytz>=2023.3